teste

+++ app.py
@@ -26,7 +26,7 @@ def carrega_imagem():
         st.image(image)
         st.success('Imagem foi carregada com sucesso')
 
-        image = np.array(image, dtype=np.float32)
+        image = np.array(image, dtype=np.float16)
         image = image / 255.0
         image = np.expand_dims(image, axis=0)
